Collapse leftover Notion admin redirects into one catch-all page.
