--- src/app/[locale]/admin/notion/analytics/page.tsx
@@ -1,8 +1,0 @@
-import { redirectLegacyNotionPage } from "../redirect-legacy";
-
-/** Legacy /admin/notion/analytics → AppFlowy analytics. */
-export default async function LegacyNotionAnalyticsPage({
-  params,
-}: Readonly<{ params: Promise<{ locale: string }> }>) {
-  await redirectLegacyNotionPage(params, "/admin/appflowy/analytics");
-}
