--- src/app/[locale]/admin/notion/page.tsx
@@ -1,8 +1,0 @@
-import { redirectLegacyNotionPage } from "./redirect-legacy";
-
-/** Legacy /admin/notion → AppFlowy CMS. */
-export default async function LegacyNotionIndexPage({
-  params,
-}: Readonly<{ params: Promise<{ locale: string }> }>) {
-  await redirectLegacyNotionPage(params, "/admin/appflowy");
-}
