--- src/app/[locale]/admin/notion/tasks/page.tsx
@@ -1,8 +1,0 @@
-import { redirectLegacyNotionPage } from "../redirect-legacy";
-
-/** Legacy /admin/notion/tasks → AppFlowy tasks. */
-export default async function LegacyNotionTasksPage({
-  params,
-}: Readonly<{ params: Promise<{ locale: string }> }>) {
-  await redirectLegacyNotionPage(params, "/admin/appflowy/tasks");
-}
